file & dir refactoring
