+++ package.json
@@ -19,6 +19,7 @@
     "eslint-plugin-playwright": "^2.1.0",
     "jest": "^29.7.0",
     "playwright-core": "^1.49.0",
+    "pre-commit": "^1.2.2",
     "prettier": "^3.3.3",
     "simple-git": "^3.27.0",
     "ts-jest": "^29.2.5",
